+++ src/main/java/uos/uos25/order/dto/request/OrdersStatusRequestDTO.java
@@ -1,4 +1,4 @@
-package uos.uos25.orders.dto.request;
+package uos.uos25.order.dto.request;
 
 import com.fasterxml.jackson.annotation.JsonCreator;
