+++ src/main/java/uos/uos25/order/dto/request/OrdersUpdateRequestDTO.java
@@ -1,4 +1,4 @@
-package uos.uos25.orders.dto.request;
+package uos.uos25.order.dto.request;
 
 import lombok.Data;
