+++ src/main/java/uos/uos25/order/entity/Orders.java
@@ -1,4 +1,4 @@
-package uos.uos25.orders.entity;
+package uos.uos25.order.entity;
 
 import jakarta.persistence.*;
 
@@ -6,8 +6,8 @@ import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import uos.uos25.common.BaseEntity;
-import uos.uos25.orders.exception.OrdersCannotBeCanceled;
-import uos.uos25.product.entity.Product;
+import uos.uos25.headQuarter.entity.Product;
+import uos.uos25.order.exception.OrdersCannotBeCanceled;
 import uos.uos25.shop.entity.Shop;
 
 @Entity
