+++ src/main/java/uos/uos25/order/exception/OrdersNotFoundException.java
@@ -1,4 +1,4 @@
-package uos.uos25.orders.exception;
+package uos.uos25.order.exception;
 
 import uos.uos25.common.error.BusinessException;
 import uos.uos25.common.error.ErrorCode;
