+++ src/main/java/uos/uos25/order/repository/OrdersRepository.java
@@ -1,11 +1,11 @@
-package uos.uos25.orders.repository;
+package uos.uos25.order.repository;
 
 import java.util.List;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import uos.uos25.orders.entity.Orders;
+import uos.uos25.order.entity.Orders;
 
 @Repository
 public interface OrdersRepository extends JpaRepository<Orders, Long> {
